# funcr: support fmt.Stringer