It seems like a logical thing to do now, though I still feel like I could argue that call-sites can CHOOSE to call `String()`, which saves epsilon overhead on every call...

Fixes #95