@thockin 2021-09-27
The cost of this is measurable:

before:

```
$ go test -bench='Funcr' ./benchmark/
goos: linux
goarch: amd64
pkg: github.com/go-logr/logr/benchmark
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkFuncrInfoOneArg-6        	 2254779	       532.3 ns/op
BenchmarkFuncrInfoSeveralArgs-6   	 1000000	      1065 ns/op
BenchmarkFuncrV0Info-6            	 1000000	      1071 ns/op
BenchmarkFuncrV9Info-6            	15084050	        78.66 ns/op
BenchmarkFuncrError-6             	  927880	      1295 ns/op
BenchmarkFuncrWithValues-6        	 7474820	       159.9 ns/op
BenchmarkFuncrWithName-6          	23154981	        51.70 ns/op
BenchmarkFuncrWithCallDepth-6     	24020500	        50.18 ns/op
```

after:
```
$ go test -bench='Funcr' ./benchmark/
goos: linux
goarch: amd64
pkg: github.com/go-logr/logr/benchmark
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkFuncrInfoOneArg-6        	 1994936	       569.9 ns/op
BenchmarkFuncrInfoSeveralArgs-6   	 1000000	      1192 ns/op
BenchmarkFuncrV0Info-6            	 1000000	      1193 ns/op
BenchmarkFuncrV9Info-6            	15369814	        78.05 ns/op
BenchmarkFuncrError-6             	  820257	      1426 ns/op
BenchmarkFuncrWithValues-6        	 7464562	       160.4 ns/op
BenchmarkFuncrWithName-6          	22276531	        54.29 ns/op
BenchmarkFuncrWithCallDepth-6     	24287416	        49.72 ns/op
```