@pohly 2021-09-27
> The cost of this is measurable, which is why I didn't do it in the first place.

That mirrors my measurements that I did for https://github.com/go-logr/zapr/pull/46. But does saving some ns matter [when a normal function call takes us](https://github.com/go-logr/zapr/pull/46#discussion_r709887125)?