@thockin 2021-09-27
sure, pushed

On Mon, Sep 27, 2021 at 12:32 PM Patrick Ohly ***@***.***>
wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In funcr/funcr.go
> <https://github.com/go-logr/logr/pull/96#discussion_r716983678>:
>
> >  		// That then gets handled below via reflection.
>  		value = v.MarshalLog()
> +	}
> +	// Handle types that want to format themselves.
> +	if v, ok := value.(fmt.Stringer); ok {
> +		value = v.String()
> +	} else if v, ok := value.(error); ok {
>
> A type switch might be slightly easier to read than if/else?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/go-logr/logr/pull/96#pullrequestreview-764668960>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABKWAVGJMULQOZQ2PRSIYH3UEDBDLANCNFSM5E27OHWQ>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>