@thockin 2021-09-27
Regarding perf - 7% is not a HUGE factor, and I think this is less surprising, so... OK I guess..